Update 1.3.6: Hidden Import and Invoice Output Out Of Project

+++ internal/service/sip_object_service.go
@@ -6,7 +6,9 @@ import (
 	"backend-v2/model"
 	"fmt"
 	"os"
+	"path/filepath"
 	"strconv"
+	"time"
 
 	"github.com/xuri/excelize/v2"
 )
@@ -16,6 +18,7 @@ type sipObjectService struct {
 	workerRepo            repository.IWorkerRepository
 	objectSupervisorsRepo repository.IObjectSupervisorsRepository
 	objectTeamsRepo       repository.IObjectTeamsRepository
+	teamRepo              repository.ITeamRepository
 }
 
 func InitSIPObjectService(
@@ -23,12 +26,14 @@ func InitSIPObjectService(
 	workerRepo repository.IWorkerRepository,
 	objectSupervisorsRepo repository.IObjectSupervisorsRepository,
 	objectTeamsRepo repository.IObjectTeamsRepository,
+	teamRepo repository.ITeamRepository,
 ) ISIPObjectService {
 	return &sipObjectService{
 		sipObjectRepo:         sipObjectRepo,
 		workerRepo:            workerRepo,
 		objectSupervisorsRepo: objectSupervisorsRepo,
 		objectTeamsRepo:       objectTeamsRepo,
+		teamRepo:              teamRepo,
 	}
 }
 
@@ -38,7 +43,7 @@ type ISIPObjectService interface {
 	Create(data dto.SIPObjectCreate) (model.SIP_Object, error)
 	Update(data dto.SIPObjectCreate) (model.SIP_Object, error)
 	Delete(id, projectID uint) error
-	TemplateFile(filepath string, projectID uint) error
+	TemplateFile(filepath string, projectID uint) (string, error)
 	Import(projectID uint, filepath string) error
 }
 
@@ -92,31 +97,44 @@ func (service *sipObjectService) Delete(id, projectID uint) error {
 	return service.sipObjectRepo.Delete(id, projectID)
 }
 
-func (service *sipObjectService) TemplateFile(filepath string, projectID uint) error {
-	f, err := excelize.OpenFile(filepath)
+func (service *sipObjectService) TemplateFile(filePath string, projectID uint) (string, error) {
+	f, err := excelize.OpenFile(filePath)
 	if err != nil {
 		f.Close()
-		return fmt.Errorf("Не смог открыть шаблонный файл: %v", err)
+		return "", fmt.Errorf("Не смог открыть шаблонный файл: %v", err)
 	}
 
 	sheetName := "Супервайзеры"
 	allSupervisors, err := service.workerRepo.GetByJobTitleInProject("Супервайзер", projectID)
 	if err != nil {
 		f.Close()
-		return fmt.Errorf("Данные супервайзеров недоступны: %v", err)
+		return "", fmt.Errorf("Данные супервайзеров недоступны: %v", err)
 	}
 
 	for index, supervisor := range allSupervisors {
 		f.SetCellValue(sheetName, "A"+fmt.Sprint(index+2), supervisor.Name)
 	}
 
-	if err := f.Save(); err != nil {
-		return fmt.Errorf("Не удалось обновить шаблон с новыми данными: %v", err)
+	allTeams, err := service.teamRepo.GetAll(projectID)
+	if err != nil {
+		f.Close()
+		return "", fmt.Errorf("Данны бригад не доступны: %v", err)
+	}
+
+	teamSheetName := "Бригады"
+	for index, team := range allTeams {
+		f.SetCellStr(teamSheetName, "A"+fmt.Sprint(index+2), team.Number)
+	}
+
+	date := time.Now()
+	tmpFilePath := filepath.Join("./pkg/excel/temp/", date.String()+" Шаблон для импорта СИП.xlsx")
+	if err := f.SaveAs(tmpFilePath); err != nil {
+		return "", fmt.Errorf("Не удалось обновить шаблон с новыми данными: %v", err)
 	}
 
 	f.Close()
 
-	return nil
+	return tmpFilePath, nil
 }
 
 func (service *sipObjectService) Import(projectID uint, filepath string) error {
@@ -128,7 +146,7 @@ func (service *sipObjectService) Import(projectID uint, filepath string) error {
 		return fmt.Errorf("Не смог открыть файл: %v", err)
 	}
 
-	sheetName := "Импорт"
+	sheetName := "СИП"
 	rows, err := f.GetRows(sheetName)
 	if err != nil {
 		f.Close()
@@ -142,9 +160,7 @@ func (service *sipObjectService) Import(projectID uint, filepath string) error {
 		return fmt.Errorf("Файл не имеет данных")
 	}
 
-	objects := []model.Object{}
-	sips := []model.SIP_Object{}
-	supervisorIDs := []uint{}
+	sips := []dto.SIPObjectImportData{}
 	index := 1
 	for len(rows) > index {
 		object := model.Object{
@@ -168,38 +184,65 @@ func (service *sipObjectService) Import(projectID uint, filepath string) error {
 			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке B%d: %v", index+1, err)
 		}
 
-		supervisorName, err := f.GetCellValue(sheetName, "C"+fmt.Sprint(index+1))
+		amountFeedersSTR, err := f.GetCellValue(sheetName, "C"+fmt.Sprint(index+1))
 		if err != nil {
 			f.Close()
 			os.Remove(filepath)
 			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке C%d: %v", index+1, err)
-		}
-
-		supervisorWorker, err := service.workerRepo.GetByName(supervisorName)
-		if err != nil {
-			f.Close()
-			os.Remove(filepath)
-			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке C%d: %v", index+1, err)
-		}
-
-		amountFeedersSTR, err := f.GetCellValue(sheetName, "D"+fmt.Sprint(index+1))
-		if err != nil {
-			f.Close()
-			os.Remove(filepath)
-			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке D%d: %v", index+1, err)
 		}
 
 		amountFeedersUINT64, err := strconv.ParseUint(amountFeedersSTR, 10, 64)
 		if err != nil {
 			f.Close()
 			os.Remove(filepath)
-			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке D%d: %v", index+1, err)
+			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке C%d: %v", index+1, err)
 		}
 		sip.AmountFeeders = uint(amountFeedersUINT64)
 
-		supervisorIDs = append(supervisorIDs, supervisorWorker.ID)
-		objects = append(objects, object)
-		sips = append(sips, sip)
+		supervisorName, err := f.GetCellValue(sheetName, "D"+fmt.Sprint(index+1))
+		if err != nil {
+			f.Close()
+			os.Remove(filepath)
+			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке D%d: %v", index+1, err)
+		}
+
+		supervisorWorker := model.Worker{}
+		if supervisorName != "" {
+			supervisorWorker, err = service.workerRepo.GetByName(supervisorName)
+			if err != nil {
+				f.Close()
+				os.Remove(filepath)
+				return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке D%d: %v", index+1, err)
+			}
+		}
+
+		teamNumber, err := f.GetCellValue(sheetName, "E"+fmt.Sprint(index+1))
+		if err != nil {
+			f.Close()
+			os.Remove(filepath)
+			return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке E%d: %v", index+1, err)
+		}
+
+    team := model.Team{}
+		if teamNumber != "" {
+			team, err = service.teamRepo.GetByNumber(teamNumber)
+			if err != nil {
+				f.Close()
+				os.Remove(filepath)
+				return fmt.Errorf("Ошибка в файле, неправильный формат данных в ячейке E%d: %v", index+1, err)
+			}
+		}
+
+		sips = append(sips, dto.SIPObjectImportData{
+      Object: object,
+      SIP: sip,
+      ObjectSupervisors: model.ObjectSupervisors{
+        SupervisorWorkerID: supervisorWorker.ID,
+      },
+      ObjectTeam: model.ObjectTeams{
+        TeamID: team.ID,
+      },
+    })
 		index++
 	}
 
@@ -211,10 +254,5 @@ func (service *sipObjectService) Import(projectID uint, filepath string) error {
 		return fmt.Errorf("Ошибка при удалении временного файла: %v", err)
 	}
 
-	_, err = service.sipObjectRepo.CreateInBatches(objects, sips, supervisorIDs)
-	if err != nil {
-		return err
-	}
-
-	return nil
+	return service.sipObjectRepo.Import(sips)
 }
